Feature/squad unit logic enhancement 🔧🛠️⚙️ (#16)
* add comments to improve unit

* add more todos, use Uint8Array for ids, instead of Float32Array

* add grid

* add improvements and simplicity in faction, squad, unit

* remove log, reduce duplicated line

* add debugging panel, add obstacles map

* remove console logs

* improve debugging

* increase cell size of the obstacles map, fix issue with random values in obstacles map

* add comment with explanation

* add run and attack & use abiltiy and run, and also fix a bunch of other bugs

* add hunters, improve grid debugging

* extratc abilities to other file

* clean up

* remove unnecessary log and comment

* cleanup, add Math.clamp polyfill, remove Utils from TS

* typo

* add searching for the enemy

* add shooting and running POC

* add running and shooting

* fix shooting

* small improvement

* divide squads by range

* add portal to squads_grid

* smal fixes and improvements

+++ src/buttons/abilities.ts
@@ -145,7 +145,7 @@ const getAbilityIconPosition = (
 export const addAbilitiesButton = (
   universeRepresentation: UniverseRepresentation,
   units: number[][],
-  squadsIds: Float32Array,
+  squadsIds: Uint32Array,
   selectAbility: (abilityId: number) => void,
 ) => {
   let squadsIdsIndex = 0
@@ -271,10 +271,10 @@ export const disableAbility = (squadId: number) => {
   ability.container.interactive = false
 }
 
-export const getAllSimilarAvailableAbilitiesIds = (): Float32Array => {
+export const getAllSimilarAvailableAbilitiesIds = (): Uint32Array => {
   const squadsIdsToCastAbility = Object.values(abilities)
     .filter(ability => ability.selected)
     .map(ability => ability.squadId)
 
-  return new Float32Array(squadsIdsToCastAbility)
+  return new Uint32Array(squadsIdsToCastAbility)
 }
